+++ source/actions/edittimesignature.cpp
@@ -45,9 +45,9 @@ void EditTimeSignature::updateFollowingTimeSignatures(
         const TimeSignature &oldTime, const TimeSignature &newTime)
 {
     Score &score = myLocation.getScore();
-    const int start_system_index = myLocation.getSystemIndex();
+    const size_t start_system_index = myLocation.getSystemIndex();
 
-    for (int i = start_system_index; i < score.getSystems().size(); ++i)
+    for (size_t i = start_system_index; i < score.getSystems().size(); ++i)
     {
         for (Barline &bar : score.getSystems()[i].getBarlines())
         {
